fixes #566
